requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
Pillow==10.0.1
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
